stop / ApplyStopsToPositionCommand: ability to specify pnl% or distance offset

+++ src/modules/Trading/Application/EventListener/RemoveBuyOrderDoublesListener.php
@@ -6,6 +6,7 @@ namespace App\Trading\Application\EventListener;
 
 use App\Bot\Domain\Repository\BuyOrderRepository;
 use App\Domain\BuyOrder\Event\BuyOrderPushedToExchange;
+use App\Helper\OutputHelper;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
 
 #[AsEventListener]
@@ -18,9 +19,17 @@ final readonly class RemoveBuyOrderDoublesListener
             return;
         }
 
+        $count = 0;
         foreach ($this->repository->getByDoublesHash($buyOrder->getDoubleHash()) as $order) {
+            if ($order->getId() === $buyOrder->getId()) {
+                continue;
+            }
+
             $this->repository->remove($order);
+            $count ++;
         }
+
+        OutputHelper::print(sprintf('martingale (%s %s): removed %d doubles', $buyOrder->getSymbol()->name(), $buyOrder->getPositionSide()->value, $count));
     }
 
     public function __construct(private BuyOrderRepository $repository)
